moved function setUpSearchForm
